Refatoring and creating migrations
