+++ server/pkg/user/model/model.go
@@ -1,0 +1,9 @@
+package model
+
+import "tournament/pkg/mid"
+
+type User struct {
+	mid.ID
+	Name    string 			`json:"name"`
+	Balance int    			`json:"balance"`
+}
